+++ R/nr-core.R
@@ -196,37 +196,6 @@ nr_replace <- function(nr, old, new) {
   invisible(.Call(replace_, nr, old, new))
 }
 
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Scale up a \code{nativeRaster} by an integer factor
-#' 
-#' @inheritParams nr_fill
-#' @param scale integer value >= 1
-#' 
-#' @return New \code{nativeRaster}
-#' 
-#' @examples
-#' nr1 <- nr_new(200, 100, 'white')
-#' nr2 <- nr_scale(nr1, 2)
-#' 
-#' @export
-#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-nr_scale <- function(nr, scale) {
-  scale <- as.integer(scale)
-  if (length(scale) != 1L || scale < 1L) {
-    stop("Invalid 'scale' value: ", scale)
-  }
-  
-  m <- matrix(1L, scale, scale)
-  mat <- as.matrix(nr)
-  dim(mat) <- rev(dim(mat))
-  res <- base::kronecker(mat, m)
-  mode(res) <- 'integer'
-  
-  class(res) <- 'nativeRaster'
-  dim(res) <- rev(dim(res))
-  res
-}
-
 
 
 
@@ -236,18 +205,6 @@ if (FALSE) {
   logo <- png::readPNG(system.file("img", "Rlogo.png", package="png"), native = TRUE)
   plot(logo, T)
 
-  big <- nr_scale(logo, 2)
-  plot(big, T)
-  
-  
-  nr <- nr_new(3, 2, 'white')
-  nr_point(nr, c(0, 1, 2, 0, 1, 2), c(0, 0, 0, 1, 1, 1), c('red', 'blue', 'green', 'black', 'hotpink', 'grey50'))
-  plot(nr, T)
-  
-  big <- nr_scale(nr, 2)
-  plot(big, T)
-  
-